+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - update to angular 17
  - add style `LabJS`
  - load `headless` pages, require SelfHelp `v6.12.1+`
+ - #10 - add `htmlTag` style
 
 ### Bugfix
  - #12 - load `textareaStyle` properly